Template functions
Former-commit-id: 24170128b788a5c1b424bff5f6c61d601999172f [formerly 6a9bd1ef4a33bcef735418346f9f91ed3c0b512d]
Former-commit-id: 923729800a0bad74645affed87a1c30b6cb6680a

+++ api/helpers/encryption/hash.go
@@ -32,7 +32,7 @@ func newSHA1Hash(n ...int) string {
 		noRandomCharacters = n[0]
 	}
 
-	randString := RandomString(noRandomCharacters, true)
+	randString := RandomString(int64(noRandomCharacters), true)
 
 	hash := sha1.New()
 	hash.Write([]byte(randString))
@@ -42,7 +42,7 @@ func newSHA1Hash(n ...int) string {
 }
 
 // randomString generates a random string of n length
-func RandomString(n int, numeric bool) string {
+func RandomString(n int64, numeric bool) string {
 	var characterRunes = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789")
 	if !numeric {
 		characterRunes = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")
